requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.3
datasets==4.0.0
debugpy==1.8.15
decorator==5.2.1
dill==0.3.8
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.7.4
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==1.6.3
google-auth-oauthlib==0.5.3
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.0
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==2.14.0
libclang==18.1.1
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
narwhals==2.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.25.2
oauthlib==3.3.1
openai==1.97.1
opt_einsum==3.4.0
packaging==25.0
pandas==1.5.3
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.9.3
sentence-transformers==5.0.0
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.47.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.14.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.4
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.14.1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
